(PC-26500)[API] perf: search offer by ean if a string with 13 digit is provided
